rework on header.tex and .chktexrc in header.tex this includes - patches of loaded packages - the new command \textTT - info on relation between amsmath and glossaries
As a consequence of the patching in header.tex, 
now also a weakness in pythontexInOut.pdf disappeared: 
now list of listings is in the table of contents. 
In the manual, the according section now in header.tex was removed. 


Also .chktex is modified: 
- added \textTT to the commands with wiped arguments. 
- added documentation from which package the entries come. 

In the manual, the patches were treated in the section on injecting header.tex
So some restructuration. 
Also changes in .chktex, i.e. adding \textTT is reflected in the Section on injection of .chktex 
Moreover, it is emphasized, that header.tex, .latexmk and .chktex have synergies. 

Also added these subjects in the section on document development under latexmk and chktex.